# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Anby Wiki API
  version: "0.1.0"
  description: |
    Anby Wiki HTTP API 权威契约（OpenAPI 3.1）。
    本文件是所有接口行为的唯一权威定义：错误模型、分页、幂等头与请求 ID 在此统一定义。
    前端客户端由 OpenAPI Generator 从本文件生成，禁止手写 URL/DTO。
    Go 侧 handler 手写实现，由 Contract Test 保证与本文件一致。
  contact:
    name: Anby Wiki
...
servers:
  - url: http://localhost:3000
    description: 本地 Web 入口（Next.js 将 /api/* 转发到 Go API）
  - url: http://localhost:8080
    description: 本地直连 Go API
...
paths:
...
  /api/v1/entities/{id}/merge:
    get:
      operationId: getEntityMerge
      summary: 读取 Entity 的最近合并记录
      description: 匿名读取源 Entity 的最近一次合并及其不可变标签、Claim 映射账本。
      tags: [knowledge]
      parameters:
        - $ref: "#/components/parameters/ReferenceTargetId"
      responses:
        "200":
          description: Entity 合并记录
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityMergeResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/Internal"
    post:
      operationId: mergeEntity
      summary: 合并重复 Entity
      description: 仅站点管理员或 system Actor 可触发。合并、审计与 entity.merged Outbox 在同一事务提交；引用修复 Proposal 由 Worker 异步幂等生成。
      tags: [knowledge]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ReferenceTargetId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeEntityRequest"
      responses:
        "200":
          description: 合并完成；同一 source/target 的重复请求返回既有结果
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EntityMergeResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
  /api/v1/entity-merges/{id}/rollback:
    post:
      operationId: rollbackEntityMerge
      summary: 补偿回滚 Entity 合并
      description: |
        仅管理员可执行。服务端在一个事务中恢复源 Entity，补偿合并产生的 Claim，
        移除只由该合并复制且未被后续修改的标签，并追加审计；任何状态漂移都会拒绝整笔回滚。
      tags: [knowledge]
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ReferenceTargetId"
      responses:
        "200":
          description: Entity 合并补偿结果；重复请求幂等返回
          headers:
            X-Request-ID:
              $ref: "#/components/headers/X-Request-ID"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RollbackEntityMergeResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/Internal"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: anby_session
      description: HttpOnly 同源服务端 session cookie；名称可由部署配置覆盖
  headers:
    X-Request-ID:
      description: 请求 ID；客户端可携带同名请求头，服务端透传或重新生成
      schema:
        type: string
  parameters:
...
    ReferenceTargetId:
      name: id
      in: path
      required: true
      description: Entity、Claim 或 Citation 稳定 ID
      schema:
        type: string
        format: uuid
...
  schemas:
...
    MergeEntityRequest:
      type: object
      additionalProperties: false
      required: [target_entity_id]
      properties:
        target_entity_id:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 1000
          default: ""
    EntityMergeLabelMapping:
      type: object
      required: [language, source_label, target_label, action, target_is_primary]
      properties:
        language:
          type: string
        source_label:
          type: string
        target_label:
          type: string
        action:
          type: string
          enum: [copied, demoted_primary, existing]
        target_is_primary:
          type: boolean
    EntityMergeClaimMapping:
      type: object
      required: [old_claim_id, new_claim_id, old_status, new_status]
      properties:
        old_claim_id:
          type: string
          format: uuid
        new_claim_id:
          type: string
          format: uuid
        old_status:
          type: string
        new_status:
          type: string
    EntityMergeResult:
      type: object
      required:
        [id, source_entity_id, target_entity_id, actor_id, status, reason,
         created_at, idempotent, label_mappings, claim_mappings]
      properties:
        id:
          type: string
          format: uuid
        source_entity_id:
          type: string
          format: uuid
        target_entity_id:
          type: string
          format: uuid
        actor_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [applied, rolled_back]
        reason:
          type: string
        created_at:
          type: string
          format: date-time
        idempotent:
          type: boolean
        label_mappings:
          type: array
          items:
            $ref: "#/components/schemas/EntityMergeLabelMapping"
        claim_mappings:
          type: array
          items:
            $ref: "#/components/schemas/EntityMergeClaimMapping"
    RollbackEntityMergeResult:
      type: object
      additionalProperties: false
      required:
        [merge_id, restored_entity_id, compensated_claim_ids,
         removed_target_labels, idempotent]
      properties:
        merge_id:
          type: string
          format: uuid
        restored_entity_id:
          type: string
          format: uuid
        compensated_claim_ids:
          type: array
          items:
            type: string
            format: uuid
        removed_target_labels:
          type: integer
          minimum: 0
        idempotent:
          type: boolean
...
    Error:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: 稳定机器可读错误码
          enum:
            - bad_request
            - unauthorized
            - forbidden
            - not_found
            - gone
            - conflict
            - stale_revision
            - validation_failed
            - rate_limited
            - internal
        message:
          type: string
          description: 面向调用者的错误描述（不泄露内部实现细节）
        request_id:
          type: string
          description: 与响应头 X-Request-ID 一致，用于日志串联
        details:
          type: object
          description: 可选结构化错误明细（如字段级校验错误）
          additionalProperties: true
...
  responses:
    BadRequest:
      description: 请求格式错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未认证
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: 已认证但无权限
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: 并发冲突（含陈旧基线、幂等键冲突）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Internal:
      description: 服务端内部错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"